Pull in older version of the networking protocol from master branch
Removed a lot of the play status protocol source
from before the API-switcharoo. Was not compatible in any way whatsoever.

In the long term, I'd like to be able to manipulate flow-network into
allowing Guice injections. Heck, Java 9 is coming out in 7 days. I'll be
abandoning Guice when Java 9 is a thing.